+++ CoreMIDI4J/src/uk/co/xfactorylibrarians/coremidi4j/CoreMidiClient.java
@@ -26,7 +26,7 @@ public class CoreMidiClient {
 
   private final int midiClientReference;
 
-  private List<CoreMidiNotification> notificationListeners = new ArrayList<CoreMidiNotification>();
+  private final List<CoreMidiNotification> notificationListeners = new ArrayList<CoreMidiNotification>();
 
   /**
    * Constructor for class