+++ CoreMIDI4J/src/uk/co/xfactorylibrarians/coremidi4j/CoreMidiNotification.java
@@ -28,6 +28,6 @@ public interface CoreMidiNotification {
    * 
    */
 
-  public void midiSystemUpdated() throws CoreMidiException;
+  void midiSystemUpdated() throws CoreMidiException;
 
 }