+++ CoreMIDI4J/src/uk/co/xfactorylibrarians/coremidi4j/CoreMidiTransmitter.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Title:        CoreMIDI4J
  * Description:  Core MIDI Device Provider for Java on OS X
  * Copyright:    Copyright (c) 2015-2016
@@ -17,6 +17,8 @@ package uk.co.xfactorylibrarians.coremidi4j;
 import javax.sound.midi.MidiDevice;
 import javax.sound.midi.MidiDeviceTransmitter;
 import javax.sound.midi.Receiver;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * CoreMidiTransmitter - used to receive data from the connected device and send it to the application.
@@ -26,7 +28,8 @@ import javax.sound.midi.Receiver;
 public class CoreMidiTransmitter implements MidiDeviceTransmitter {
 
   private final CoreMidiSource device;
-  private Receiver receiver = null;
+  private final AtomicReference<Receiver> receiver = new AtomicReference<>();
+  private final AtomicBoolean closed = new AtomicBoolean(false);
 
   /**
    * CoreMidiTransmitter constructor
@@ -35,7 +38,7 @@ public class CoreMidiTransmitter implements MidiDeviceTransmitter {
    * 
    */
 
-  public CoreMidiTransmitter(final CoreMidiSource device) {
+  CoreMidiTransmitter(final CoreMidiSource device) {
 
     this.device = device;
 
@@ -46,14 +49,14 @@ public class CoreMidiTransmitter implements MidiDeviceTransmitter {
    * 
    * @see javax.sound.midi.Transmitter#setReceiver(javax.sound.midi.Receiver)
    * 
-   * @param receiver	The receiver to set
+   * @param receiver	The receiver to set, replacing any previous value
    * 
    */
 
   @Override
   public void setReceiver(Receiver receiver) {
 
-    this.receiver = receiver;
+    this.receiver.set(receiver);
 
   }
 
@@ -69,12 +72,12 @@ public class CoreMidiTransmitter implements MidiDeviceTransmitter {
   @Override
   public Receiver getReceiver() {
 
-    return receiver;
+    return receiver.get();
 
   }
 
   /**
-   * Closes this transmitter
+   * Closes this transmitter, causing it to no longer send MIDI events from its source
    * 
    * @see javax.sound.midi.Transmitter#close()
    * 
@@ -83,7 +86,11 @@ public class CoreMidiTransmitter implements MidiDeviceTransmitter {
   @Override
   public void close() {
 
-    receiver = null;
+    if (closed.compareAndSet(false, true)) {
+
+      device.transmitterClosed(this);
+
+    }
 
   }